systemd: Reorder dependencies and config options
This is mostly a readability commit.

It essentially doesn’t change anything in the expression, except that
it splits the expression parameters into logical groups and reorders
the configuration options to split them into logical groups as well
and to have them in the order that more closely resembles that of
the upstream documentation and configuration scripts.

The only two changes to the semantics of the expression are:

* `withSelinux` is removed and replaced with `selinux != null`.
* `lz4` is not forced-enabled, instead the build system will detect
  whether it is available or not.